+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this application adheres to [Semantic Versioning](https://semver.org/spec/v2
 
 ## [Unreleased]
 
+### Fixed
+- Fiat text field on Send screen is disabled if fetching exchange rate fails
+
+### Changed
+- When entering amount in USD in the Send or Request ZEC flow, we floor the Zatoshi amount automatically to the nearest 5000 Zatoshi to prevent creating unspendable dust notes in your wallet.
+- Integrations dialog screen now displays a disclaimer
+- Primary & secondary button order now follows UX best practices
+- Receive screen design now aligns better with UX after home navigation changes
+- Copy updated in a few places
+
 ## [2.0.1 (941)] - 2025-04-29
 
 ### Fixed